requirements.txt
discord.py==2.0.0a3428+gecf239d2
gspread==4.0.1
requests==2.22.0
oauth2client==4.1.3